[IMP] ignore HttpCase test without --odoo-http option
