+++ tox.ini
@@ -4,9 +4,11 @@ envlist =
     py39-odoo{14,15}
     py310-odoo{14,15,16}
     py311-odoo{15,16,17}
-    py312-odoo{16,17,18}
+    py312-odoo{16,17,18,19}
     py313-odoo{17,18,19}
-    py{38,39,310,311,312,313}-unit
+    # Odoo does not support python 3.14 yet
+    # py314-odoo{18,19}
+    py{38,39,310,311,312,313,314}-unit
 
 [testenv]
 usedevelop = true
@@ -43,7 +45,7 @@ commands_pre =
     # Odoo 18
     py{312,313,314}-odoo18: pip install lxml>=5.2.1 lxml_html_clean Werkzeug==3.0.1 freezegun
     # Odoo 19
-    py{313,314}-odoo19: pip install lxml>=5.2.1 lxml_html_clean Werkzeug==3.0.1 PyPDF==5.4.0  freezegun
+    py{312,313,314}-odoo19: pip install lxml>=5.2.1 lxml_html_clean Werkzeug==3.0.1 PyPDF==5.4.0 freezegun
     odoo14: pip install https://github.com/OCA/OCB/archive/14.0.tar.gz
     odoo15: pip install https://github.com/OCA/OCB/archive/15.0.tar.gz
     odoo16: pip install https://github.com/OCA/OCB/archive/16.0.tar.gz
@@ -54,5 +56,8 @@ commands_pre =
     unit: pip install tests/mock/odoo
 
 commands =
+    # launch odoo http server using xdist in different processes
     !unit: pytest -n 2 --odoo-database {envname} --odoo-http -vvv --cov-append --odoo-addons-path={toxinidir}/tests_integration/odoo/addons --cov=pytest_odoo --cov-report=xml --cov-report=term-missing {toxinidir}/tests_integration/odoo/addons/pytest_odoo_test_module/ {posargs}
+    # http test case should be skipped without xdist
+    !unit: pytest --odoo-database {envname} -vvv --cov-append --odoo-addons-path={toxinidir}/tests_integration/odoo/addons --cov=pytest_odoo --cov-report=xml --cov-report=term-missing {toxinidir}/tests_integration/odoo/addons/pytest_odoo_test_module/ {posargs}
     unit: pytest --cov-append --cov=pytest_odoo --cov-report=xml --cov-report=term-missing tests/ {posargs}
